set content-type header in server response
The server will now set the content-type header in the response by default, when it is known. I.e.:
- when the server receives a struct response from the API impl that it serializes to JSON, it sets the content type to `application/json`
- when the API impl returns a string, it passes a `text/plain` as response type.

The accept header sent by the client would override the default behavior though. The server applies some simple heuristics to match accepted mime type:
- for JSON responses, a mime of the form `*/json` is chosen if available
- for text responses, a mime of the form `text/*` is chosen if available

If the default heuristics do not suit an implementation, it could always return a `Response` type with the exact content-type header set.
If no content type could be matched, the server still passes the response forward in the processing chain, but without a content-type header. This optimistic behavior suits situations where something else in the processing chain could add the header, or the client could be lenient about the content-type.