allow escaping of path params to be optional (#58)
Certain OpenAPI specs depend on the path params to be not escaped (e.g. OpenPolicyAgent specifications).

This will allow the escaping of path parameters to be specified at request context level (with default as `true`), and also allows overriding at the client instance level (defaults to no overridiing). The existing behavior of always escaping is retained as the default.