feat: helpers for serverside validations
This adds validation method templates that can be used by `openapi-generator` to generate better serverside validation code.

All client and server code have been updated using a branch of `openapi-generator` that does that. A PR with that to `openapi-generator` repo will follow. A few test methods had to be changed to conform to the additional validations.